Add GitLab CI, repo-wide lint config, and fix findings
CI follows the same shape as cove: lint/test/build stages, base images pulled
through GitLab's dependency proxy, and a dind image build that passes the same
prefix in as BASE_REGISTRY so the Dockerfile's FROM lines use the cache too.

No frontend lint job — the SPA is plain JavaScript with nothing configured, so
build:frontend is the real compile-time check. It asserts index.html was
actually emitted, since a silently empty build would ship an empty SPA.

Lint config lives in ruff.toml at the repo root rather than backend/. Ruff
resolves config per file by searching upward, so a config under backend/ never
reached cli/ — which was being linted with default rules. `src` is declared for
the same reason: without it `app` resolves as third-party from the root and
import sorting flips depending on the invocation directory.

Configured flake8-bugbear's extend-immutable-calls for FastAPI: B008 flagged
118 uses of the `Depends(...)` idiom, which is documented FastAPI usage rather
than a mutable-default bug.

Fixes for the genuine findings, including one real bug: `minireg login
--scopes` parsed the flag and never sent it, so a request for publish/admin was
silently dropped. Scopes now travel to the registry and pre-select on the
approval screen; the server still caps them at the approver's own authority.

Enums move to StrEnum (py312). Verified against the live database first —
SQLAlchemy persists by .name, which is unchanged, and nothing interpolates a
bare enum.

+++ backend/app/services/osv.py
@@ -105,7 +105,7 @@ def _vector_parts(vector: str) -> dict[str, str]:
 
 def _round_up(value: float) -> float:
     """CVSS 3.1 Appendix A roundup, done in integer space to dodge float error."""
-    integer = int(round(value * 100000))
+    integer = round(value * 100000)
     if integer % 10000 == 0:
         return integer / 100000.0
     return (int(integer / 10000) + 1) / 10.0
@@ -128,7 +128,10 @@ def _cvss3_base(vector: str) -> float | None:
         return None
 
     iss = 1 - ((1 - conf) * (1 - integ) * (1 - avail))
-    if scope_changed:
+    # Kept as a branch rather than a ternary: these are the two impact
+    # sub-formulas from the CVSS 3.1 specification, and side by side they can
+    # be checked against it line for line.
+    if scope_changed:  # noqa: SIM108
         impact = 7.52 * (iss - 0.029) - 3.25 * ((iss - 0.02) ** 15)
     else:
         impact = 6.42 * iss
@@ -621,7 +624,7 @@ async def inline_scan(
     except (TimeoutError, asyncio.CancelledError):
         log.info("inline OSV scan timed out for %s@%s", package_name, version)
         return ScanResult(None, package_name, version, scanned=False)
-    except Exception as exc:  # noqa: BLE001
+    except Exception as exc:
         log.warning("inline OSV scan failed for %s@%s: %s", package_name, version, exc)
         return ScanResult(None, package_name, version, scanned=False)
     return results.get(
